included method to date back 30 days

+++ Five/CoreData/Entities/Day+CoreDataProperties.swift
@@ -18,15 +18,19 @@ extension Day {
     }
 
     @NSManaged public var createdAt: NSDate?
-    @NSManaged public var taskLimit: Int64
+    // Total Task limit for the day
+    @NSManaged public var taskLimit: Int16
     @NSManaged public var year: Int16
     @NSManaged public var month: Int16
     
-    //Represents the date of the month not the day of the week
+    // Represents the date of the month not the day of the week
     @NSManaged public var day: Int16
     
     @NSManaged public var totalCompleted: Int16
+    
+    // Total tasks the user has manually input
     @NSManaged public var totalTasks: Int16
+    
     @NSManaged public var dayToTask: NSSet?
 
 }
